✚ Added basic verb ingestion
This includes reflexivity of a verb, denoted as either no, conditional,
and mandatory.  The existing prompts handle the conjugations for no, and
mandatory just fine.  Splitting into multiple conjugations are going to
be required later for the conditional cases.  For now, they will be
treated as if they were mandatory cases.

Splitting of the main prompt has been started, and should continue, as
the existing prompt is getting too unreliable - it often groups the
tenses into separate objects at the top level, instead of under a
'tense' group.

The added verbs ORM bits can probably be combined with the existing
conjugation bits into a single 'models' file.