🇫🇷 Basic and working sentence prompt generation
Not using direct, indirect pronouns, or negations yet, but otherwise working smoothly.

ChatGTP seems quite able to fill in details after what was meant to be quizzed on, so I'll leave that in for now!

The ORM mapping (meant for convinience) seems not needed, so I'll likely remove that later.