🇫🇷 Added indirect pronoun support
Feature is still turned on by default, like direct objects.  This feature for now includes the adverbial pronouns y, and en.  It can be decided later if that should be a separate feature as well.  For now it doesn't seem necessary.

Refined the prompts further.  It seems like almost all the questions are being generated correctly now, with only a few misformattings here and there.