➕ Add support for batched problem generation
This does not succeed 100% of the time due to ChatGTP rate limiting.  It seems to work about 95% of the time up to the introductory rate limiting allowed at the minimum account tier.

Exponential backoff on requests was added to mitigate that in the short term.  Batching of sentence requests for forming problems will help further in the future.