💪 Added better negation support
Made it displayable in the sentence generation output.

This effort exposed a lot of other problems generating sentences.  First of all, by not properly serializing the AI response into proper objects, the truthiness of python takes all responses as true.  This is whey COD/COI was always incorrectly showing as present.  Additionally, 'random' as a word was appearing everywhere random was selected as a setting.  Both of those require re-evaluation of how enums are being used everywhere.

The prompts are still generally fine as long as proper objects can be made to replace the string JSON.