💪 Sentence fetching from database working
Turns out painful grinding of the docs was adequate to solve this immediate problem.

This required an overhaul to now the models were being automaped.  This overhaul will be applied to the rest of the tables in subsequent commits.

Sentence fetching only works on infinitives and correctness at this point, but adding the remaining features shouldn't prove to be much of a problem later.