💪 Add error correction, and increase precision of sentence prompts
A lot of additional specificity was needed in the prompt components, which allowed for a simple correctness check in the end.

Generated sentence correctness was tested to 97% correct (with about 10% having minor ideomatic or context errors - fine for quizzing), and the 3% having moderate to major ones.  A feedback mechanism later can remove those.

This change has reinforced the need for a massive cleanup of the prompt generators, as they have grown out of control and difficult to manage.