🐞 Fixed a bug where an empty string could be returned in the negation field of a sentence
As negation is always set to 'random' in the lack of any choice, and the returned negation is whatever the sentence generator makes, it is a choice to force it to any other value.

By forcing the presence of 'ne' in all random cases in the prompt, an empty string can be reteurned.