🗄️ Split database scripts, and add cloud readme
Database creation is not necessarily automatic on cloud instances, however it is not something that is needed for local testing, so it should be separate.

Also, a readme just for myself so I don't forget the steps needed to provision and configure one of these on RDS, and make sure my private domain points at it properly as this will remain a manual process for now.